refactor: remove basic code
