--- src/hash/files/fileToCalculateHashFor.txt
@@ -1,1 +1,0 @@
-Calculate hash for me!
