+++ src/index.js
@@ -12,3 +12,8 @@ import { welcomeUser } from './greeting/welcomeUser.js';
 
 welcomeUser();
 
+function getPath() {
+    console.log(`\nYou are currently in ${currentPath.path}`);
+}
+
+getPath();
